distributed_system/sentiment_analyzer/requirements.txt
kafka-python
transformers
torch